change map structure
