+++ src/routes/map/constants/layout.constants.ts
@@ -1,0 +1,18 @@
+export const COMPLETE_SCREEN = 24;
+export const MEDIUM_SCREEN = 15;
+export const EMPTY_SCREEN = 0;
+
+export const MEDIUM_SCREEN_LEFT = 15;
+export const MEDIUM_SCREEN_RIGHT = 9;
+
+export const MAP_RESIZABLE_TRANSITION = 0.7;
+export const PROBLEMS_TRIGGER = 'problems';
+
+export const PROJECTS_LINE = 'mhfd_projects';
+
+export const PROJECTS_MAP_STYLES = {
+  name: 'projects',
+  tiles: [
+    PROJECTS_LINE
+  ]
+};
