+++ src/routes/map/constants/tabs.constants.ts
@@ -1,0 +1,5 @@
+export const PROBLEMS_TRIGGER = 'problems';
+export const PROJECTS_TRIGGER = 'projects';
+
+export const FILTER_PROBLEMS_TRIGGER = 'Problems';
+export const FILTER_PROJECTS_TRIGGER = 'Projects';
