+++ src/routes/map/index.ts
@@ -1,0 +1,3 @@
+import MapLayout from './components/MapLayout';
+
+export default MapLayout;
